Modify and add question prompts
