+++ rdt-open-reader/src/covidIndonesia/java/io/ona/rdt/activity/CovidLoginActivity.java
@@ -10,13 +10,11 @@ import io.ona.rdt.interactor.RDTJsonFormInteractor;
 import io.ona.rdt.widget.CovidRDTLabelFactory;
 import io.ona.rdt.widget.GoogleCovidRDTBarcodeFactory;
 import io.ona.rdt.widget.OneScanCovidRDTBarcodeFactory;
-import io.ona.rdt.widget.UWCovidRDTCaptureFactory;
 
 import static com.vijay.jsonwizard.constants.JsonFormConstants.LABEL;
 import static io.ona.rdt.util.CovidConstants.Table.COVID_PATIENTS;
 import static io.ona.rdt.util.CovidConstants.Widget.GOOGLE_COVID_BARCODE_READER;
 import static io.ona.rdt.util.CovidConstants.Widget.ONE_SCAN_COVID_BARCODE_READER;
-import static io.ona.rdt.util.CovidConstants.Widget.UW_COVID_RDT_CAPTURE;
 
 /**
  * Created by Vincent Karuri on 19/06/2020
@@ -31,7 +29,6 @@ public class CovidLoginActivity extends LoginActivity {
 
     private void registerWidgets() {
         Map<String, FormWidgetFactory> widgetFactoryMap = RDTJsonFormInteractor.getInstance().map;
-        widgetFactoryMap.put(UW_COVID_RDT_CAPTURE, new UWCovidRDTCaptureFactory());
         widgetFactoryMap.put(GOOGLE_COVID_BARCODE_READER, new GoogleCovidRDTBarcodeFactory());
         widgetFactoryMap.put(ONE_SCAN_COVID_BARCODE_READER, new OneScanCovidRDTBarcodeFactory());
         widgetFactoryMap.put(LABEL, new CovidRDTLabelFactory());
