+++ rdt-open-reader/src/covidIndonesia/java/io/ona/rdt/util/CovidConstants.java
@@ -6,7 +6,6 @@ package io.ona.rdt.util;
 public interface CovidConstants extends Constants {
 
     interface Widget {
-        String UW_COVID_RDT_CAPTURE = "uw_covid_rdt_capture";
         String GOOGLE_COVID_BARCODE_READER = "google_covid_barcode_reader";
         String ONE_SCAN_COVID_BARCODE_READER = "one_scan_covid_barcode_reader";
     }
@@ -70,8 +69,6 @@ public interface CovidConstants extends Constants {
         String COVID_SAMPLE_ID = "covid_sample_id";
         String PATIENT_FIRST_NAME = "patient_first_name";
         String PATIENT_LAST_NAME = "patient_last_name";
-        String NATIONAL_ID_NUMBER = "national_id_number";
-        String PATIENT_AGE = "patient_age";
         String PATIENT_SEX = "patient_sex";
         String LAST_KNOWN_LOCATION = "last_known_location";
         String DRIVERS_LICENSE_NUMBER = "drivers_license_number";
