--- rdt-open-reader/src/covidIndonesia/java/io/ona/rdt/widget/UWCovidRDTCaptureFactory.java
@@ -1,15 +1,0 @@
-package io.ona.rdt.widget;
-
-import org.json.JSONException;
-
-import io.ona.rdt.domain.ParcelableImageMetadata;
-
-/**
- * Created by Vincent Karuri on 17/06/2020
- */
-public class UWCovidRDTCaptureFactory extends UWRDTCaptureFactory {
-    @Override
-    protected void populateRelevantFields(ParcelableImageMetadata parcelableImageMetadata) throws JSONException {
-
-    }
-}
