+++ rdt-open-reader/src/covidIndonesia/resources/rdt-capture-form.properties
@@ -4,21 +4,24 @@ rdt-capture-form-original.step4.health_facility_type.options.health_facility.tex
 rdt-capture-form-original.step4.health_facility_name_lbl.text = Name of facility?
 rdt-capture-form-original.step6.manual_expiration_date.v_required.err = Please specify the expiration date
 rdt-capture-form-original.step2.lbl_rdt_expired.text = RDT is expired!
-rdt-capture-form-original.step6.lbl_rdt_manufacturer.text = Enter manufacturer name
 rdt-capture-form-original.step2.next_label = OK
 rdt-capture-form-original.step12.title = Handle RDT
 rdt-capture-form-original.step2.title = RDT
 rdt-capture-form-original.step6.title = Record RDT information
+rdt-capture-form-original.step8.temp_sensor.text = Sensor triggered?
 rdt-capture-form-original.step8.qr_code_reader.v_required.err = This field is required
+rdt-capture-form-original.step10.rdt_capture_bottom_line_reading.text = Bottom line reading?
 rdt-capture-form-original.step11.chw_result.options.invalid.text = invalid
 rdt-capture-form-original.step6.rdt_type.v_required.err = This field is required
 rdt-capture-form-original.step6.other_rdt_type_lbl.text = Other RDT type
+rdt-capture-form-original.step6.rdt_id.text = RDT ID?
 rdt-capture-form-original.step4.health_facility_type.v_required.err = This field is required
 rdt-capture-form-original.step11.antibody_type.options.igm_positive.text = IgM
-rdt-capture-form-original.step6.rdt_manufacturer.v_required.err = 
+rdt-capture-form-original.step6.rdt_manufacturer.v_required.err =
 rdt-capture-form-original.step2.lbl_use_new_test.text = Please use a new test.
-rdt-capture-form-original.step8.qr_code_reader.hint = 
+rdt-capture-form-original.step8.qr_code_reader.hint =
 rdt-capture-form-original.step12.next_label = CONTINUE
+rdt-capture-form-original.step8.lot_no.text = Lot number?
 rdt-capture-form-original.step7.next_label = CONTINUE
 rdt-capture-form-original.step4.health_facility_type.options.clinic.text = Clinic
 rdt-capture-form-original.step5.lbl_scan_barcode.text = Scan barcode
@@ -26,38 +29,46 @@ rdt-capture-form-original.step4.health_facility_type.label = Referral Health Fac
 rdt-capture-form-original.step9.lbl_rdt_instructions.text = Conduct the RDT according to manufacturing guidelines
 rdt-capture-form-original.step3.title = RDT
 rdt-capture-form-original.step13.title = Test complete
-rdt-capture-form-original.step6.lbl_rdt_batch_id.text = Enter RDT batch ID
 rdt-capture-form-original.step11.antibody_type.options.igg_positive.text = IgG
 rdt-capture-form-original.step4.health_facility_type.options.other.text = Other
 rdt-capture-form-original.step11.chw_result.options.positive.text = positive
 rdt-capture-form-original.step6.next_label = CONTINUE
 rdt-capture-form-original.step11.chw_result.v_required.err = This field is required
 rdt-capture-form-original.step11.chw_result.label = RDT result?
+rdt-capture-form-original.step8.gtin.text = GTIN?
+rdt-capture-form-original.step10.rdt_type.text = RDT Type?
 rdt-capture-form-original.step6.manual_expiration_date.hint = RDT expiration date
 rdt-capture-form-original.step7.lbl_handle_rdt.text = Record ID on RDT
 rdt-capture-form-original.step6.rdt_type.options.single_line.text = IgG/IgM Single Line Combo
 rdt-capture-form-original.step11.antibody_type.v_required.err = Please specify if positive for pv, pf or both
+rdt-capture-form-original.step6.rdt_batch_id_lbl.text = Enter RDT batch ID
 rdt-capture-form-original.step7.title = RDT
 rdt-capture-form-original.step13.submit_label = SAVE AND EXIT
 rdt-capture-form-original.step6.rdt_type.label = RDT type?
+rdt-capture-form-original.step10.rdt_capture_middle_line_reading.text = Middle line reading?
 rdt-capture-form-original.step11.next_label = CONTINUE
 rdt-capture-form-original.step12.lbl_handle_rdt.text = Store or dispose of RDT appropriately according to established protocol
-rdt-capture-form-original.step8.qr_code_reader.scanButtonText = 
+rdt-capture-form-original.step8.qr_code_reader.scanButtonText =
+rdt-capture-form-original.step10.rdt_capture_top_line_reading.text = Control reading?
 rdt-capture-form-original.step1.title = Find location
-rdt-capture-form-original.step6.other_rdt_type.v_required.err = 
+rdt-capture-form-original.step6.other_rdt_type.v_required.err =
 rdt-capture-form-original.step9.next_label = CAPTURE RDT
+rdt-capture-form-original.step6.rdt_manufacturer_lbl.text = Enter manufacturer name
 rdt-capture-form-original.step10.title = RDT
 rdt-capture-form-original.step4.title = RDT
 rdt-capture-form-original.step6.rdt_type.options.other.text = Other
 rdt-capture-form-original.step11.chw_result.options.negative.text = negative
 rdt-capture-form-original.step4.next_label = CONTINUE
 rdt-capture-form-original.step4.health_facility_type.options.hospital.text = Hospital
-rdt-capture-form-original.step6.rdt_batch_id.v_required.err = 
+rdt-capture-form-original.step6.rdt_batch_id.v_required.err =
+rdt-capture-form-original.step6.manual_expiration_date.text = Expiration date?
+rdt-capture-form-original.step8.unique_id.text = RDT ID?
+rdt-capture-form-original.step8.exp_date.text = Expiration date?
 rdt-capture-form-original.step11.title = Record result
 rdt-capture-form-original.step11.antibody_type.label = Antibody:
 rdt-capture-form-original.step6.rdt_type.options.double_line.text = IgG/IgM Double Line
 rdt-capture-form-original.step5.lbl_which_rdt.text = Which RDT?
-rdt-capture-form-original.step4.health_facility_name.v_required.err = 
+rdt-capture-form-original.step4.health_facility_name.v_required.err =
 rdt-capture-form-original.step5.title = RDT
 rdt-capture-form-original.step8.qr_code_reader.v_numeric.err = Invalid ID, value must be numeric
 rdt-capture-form-original.step13.lbl_test_complete.text = Diagnostic test collection for the patient has concluded. Using approved protocols, inform the patient of their results and followup procedures
