+++ rdt-open-reader/src/covidIndonesia/resources/sample-collection-form.properties
@@ -1,13 +1,13 @@
 sample-collection-form-original.step10.sputum_color.label = Color of sputum?
 sample-collection-form-original.step3.saliva_collection_kit_brand_lbl.text = Saliva collection kit brand?
 sample-collection-form-original.step3.name_of_health_facility.options.dinas_kesehatan.text = Dinas Kesehatan
-sample-collection-form-original.step3.other_health_facility_type.v_required.err = 
+sample-collection-form-original.step3.other_health_facility_type.v_required.err =
 sample-collection-form-original.step3.name_of_health_facility.options.alamat_fasyankes.text = Alamat Fasyankes
 sample-collection-form-original.step8.title = Store specimen for courier
-sample-collection-form-original.step3.health_facility_type.label = Type of health facility
+sample-collection-form-original.step3.health_facility_type.label = Type of health facility?
 sample-collection-form-original.step1.next_label = CONTINUE
 sample-collection-form-original.step3.sample_type.options.saliva.text = Saliva
-sample-collection-form-original.step6.qr_code_reader.v_numeric.err = 
+sample-collection-form-original.step6.qr_code_reader.v_numeric.err =
 sample-collection-form-original.step3.gsi_sample_number_lbl.text = GSI sample ID
 sample-collection-form-original.step7.next_label = CONTINUE
 sample-collection-form-original.step10.next_label = CONTINUE
@@ -34,12 +34,12 @@ sample-collection-form-original.step3.patient_health_status.options.recovered.te
 sample-collection-form-original.step3.name_of_health_facility.label = Name of Health Facility?
 sample-collection-form-original.step10.quantity_of_serum_lbl.text = Quantity of serum in mL?
 sample-collection-form-original.step3.sample_type.options.sputum.text = Sputum
-sample-collection-form-original.step6.qr_code_reader.v_required.err = 
+sample-collection-form-original.step6.qr_code_reader.v_required.err =
 sample-collection-form-original.step3.hospital_sample_number_lbl.text = Hospital sample ID
 sample-collection-form-original.step10.sputum_color.options.clear.text = Clear
 sample-collection-form-original.step3.name_of_health_facility.options.jenis_fasyankes.text = Jenis Fasyankes
 sample-collection-form-original.step3.next_label = CONTINUE
-sample-collection-form-original.step3.other_health_facility_type_name.v_required.err = 
+sample-collection-form-original.step3.other_health_facility_type_name.v_required.err =
 sample-collection-form-original.step3.name_of_health_facility.v_required.err = This field is required
 sample-collection-form-original.step3.swab_type.options.nasopharyngeal.text = Nasopharyngeal
 sample-collection-form-original.step3.doctor_phone_number.v_numeric.err = Value must be numeric
@@ -63,7 +63,7 @@ sample-collection-form-original.step3.vtm_brand.v_required.err = This field is r
 sample-collection-form-original.step2.lbl_collect_new_sample.text = Please collect another sample.
 sample-collection-form-original.step3.sample_type.options.other.text = Other
 sample-collection-form-original.step3.vtm_brand.options.zeesan.text = Zeesan
-sample-collection-form-original.step3.doctor_name.v_required.err = 
+sample-collection-form-original.step3.doctor_name.v_required.err =
 sample-collection-form-original.step3.health_facility_type.options.other.text = Other
 sample-collection-form-original.step7.title = Sample collection
 sample-collection-form-original.step3.other_health_facility_type_lbl.text = Other facility type
@@ -74,7 +74,7 @@ sample-collection-form-original.step3.vtm_expiration_date.hint = VTM expiration 
 sample-collection-form-original.step5.lbl_affix_respiratory_specimen_label.text = Manually affix label
 sample-collection-form-original.step3.vtm_brand.options.i3l.text = i3L
 sample-collection-form-original.step3.health_facility_type.options.hospital.text = Hospital
-sample-collection-form-original.step6.qr_code_reader.hint = 
+sample-collection-form-original.step6.qr_code_reader.hint =
 sample-collection-form-original.step3.sample_type.options.swab.text = Swab
 sample-collection-form-original.step3.serum_type.options.venus.text = Venus
 sample-collection-form-original.step9.submit_label = SAVE AND EXIT
@@ -85,10 +85,10 @@ sample-collection-form-original.step3.health_facility_type.options.primary_healt
 sample-collection-form-original.step3.patient_health_status.options.ill.text = Ill
 sample-collection-form-original.step9.title = Test complete
 sample-collection-form-original.step3.health_facility_type.options.community_health_facility.text = Community health facility
-sample-collection-form-original.step3.other_vtm_brand.v_required.err = 
+sample-collection-form-original.step3.other_vtm_brand.v_required.err =
 sample-collection-form-original.step3.other_health_facility_type_name_lbl.text = Other health facility
 sample-collection-form-original.step8.lbl_store_specimen_for_courier.text = Store the respiratory sample appropriately until they have been picked up by the courier
-sample-collection-form-original.step6.qr_code_reader.scanButtonText = 
+sample-collection-form-original.step6.qr_code_reader.scanButtonText =
 sample-collection-form-original.step3.other_sample_type.v_required.err = 
 sample-collection-form-original.step3.vtm_brand.options.fkui.text = FKUI
 sample-collection-form-original.step2.title = Sample collection
@@ -98,5 +98,5 @@ sample-collection-form-original.step3.sputum_collection_tube_brand_lbl.text = Sp
 sample-collection-form-original.step3.vtm_brand.options.nest.text = NEST
 sample-collection-form-original.step3.vtm_brand.label = Brand of VTM?
 sample-collection-form-original.step3.vtm_brand.options.lmk.text = LMK
-sample-collection-form-original.step3.doctor_phone_number.v_required.err = 
+sample-collection-form-original.step3.doctor_phone_number.v_required.err =
 sample-collection-form-original.step2.lbl_sample_expired.text = Sample is expired!
