+++ rdt-open-reader/src/main/assets/ec_client_fields.json
@@ -328,7 +328,7 @@
           "type": "Event",
           "json_mapping": {
             "field": "obs.fieldCode",
-            "concept": "rdt_capture_top_line_result"
+            "concept": "rdt_capture_top_line_reading"
           }
         },
         {
@@ -336,7 +336,7 @@
           "type": "Event",
           "json_mapping": {
             "field": "obs.fieldCode",
-            "concept": "rdt_capture_middle_line_result"
+            "concept": "rdt_capture_middle_line_reading"
           }
         },
         {
@@ -344,7 +344,7 @@
           "type": "Event",
           "json_mapping": {
             "field": "obs.fieldCode",
-            "concept": "rdt_capture_bottom_line_result"
+            "concept": "rdt_capture_bottom_line_reading"
           }
         },
         {
