+++ rdt-open-reader/src/main/java/io/ona/rdt/interactor/RDTJsonFormInteractor.java
@@ -2,14 +2,12 @@ package io.ona.rdt.interactor;
 
 import com.vijay.jsonwizard.interactors.JsonFormInteractor;
 
-import org.json.JSONObject;
-
 import io.ona.rdt.widget.MalariaRDTBarcodeFactory;
 import io.ona.rdt.widget.RDTCountdownTimerFactory;
 import io.ona.rdt.widget.RDTExpirationDateReaderFactory;
 import io.ona.rdt.widget.RDTGpsFactory;
 import io.ona.rdt.widget.RDTLabelFactory;
-import io.ona.rdt.widget.UWMalariaRDTCaptureFactory;
+import io.ona.rdt.widget.UWRDTCaptureFactory;
 
 import static com.vijay.jsonwizard.constants.JsonFormConstants.BARCODE;
 import static com.vijay.jsonwizard.constants.JsonFormConstants.COUNTDOWN_TIMER;
@@ -35,7 +33,7 @@ public class RDTJsonFormInteractor extends JsonFormInteractor {
         map.put(BARCODE, new MalariaRDTBarcodeFactory());
         map.put(LABEL, new RDTLabelFactory());
         map.put(EXPIRATION_DATE_CAPTURE, new RDTExpirationDateReaderFactory());
-        map.put(RDT_CAPTURE, new UWMalariaRDTCaptureFactory());
+        map.put(RDT_CAPTURE, new UWRDTCaptureFactory());
         map.put(COUNTDOWN_TIMER, new RDTCountdownTimerFactory());
         map.put(GPS, new RDTGpsFactory());
     }
